Do not convert 64-bit fields into number values
The number type can reliably hold only 53 bits, so do not attempt to
convert 64-bit column values into numbers. The caller can provide a
custom `cast` function to convert these into `BigInt` values if the
runtime supports it.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/MAX_SAFE_INTEGER

+++ __tests__/index.test.ts
@@ -45,14 +45,14 @@ describe('execute', () => {
     const mockResponse = {
       session: mockSession,
       result: {
-        fields: [{ name: ':vtg1', type: 'INT64' }],
+        fields: [{ name: ':vtg1', type: 'INT32' }],
         rows: [{ lengths: ['1'], values: 'MQ==' }]
       }
     }
 
     const want: ExecutedQuery = {
       headers: [':vtg1'],
-      types: { ':vtg1': 'INT64' },
+      types: { ':vtg1': 'INT32' },
       rows: [{ ':vtg1': 1 }],
       error: null,
       size: 1,
@@ -248,7 +248,7 @@ describe('execute', () => {
     const mockResponse = {
       session: null,
       result: {
-        fields: [{ name: ':vtg1', type: 'INT64' }],
+        fields: [{ name: ':vtg1', type: 'INT32' }],
         rows: [{ lengths: ['1'], values: 'MQ==' }]
       }
     }
@@ -256,7 +256,7 @@ describe('execute', () => {
     const want: ExecutedQuery = {
       headers: [':vtg1'],
       rows: [{ ':vtg1': 1 }],
-      types: { ':vtg1': 'INT64' },
+      types: { ':vtg1': 'INT32' },
       size: 1,
       error: null,
       insertId: null,
@@ -282,14 +282,14 @@ describe('execute', () => {
     const mockResponse = {
       session: null,
       result: {
-        fields: [{ name: ':vtg1', type: 'INT64' }],
+        fields: [{ name: ':vtg1', type: 'INT32' }],
         rows: [{ lengths: ['1'], values: 'MQ==' }]
       }
     }
 
     const want: ExecutedQuery = {
       headers: [':vtg1'],
-      types: { ':vtg1': 'INT64' },
+      types: { ':vtg1': 'INT32' },
       rows: [{ ':vtg1': 1 }],
       size: 1,
       error: null,
